RFR-Inpainting/requirements.txt
Python >= 3.5

PyTorch >= 1.0.0

Opencv2 ==3.4.1

Scipy == 1.1.0

Numpy == 1.14.3

Scikit-image (skimage) == 0.13.1

Oneflow_cu == 102

PIL == 4.1.1
